# backend/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
    ErrorResponseLower:
      type: object
      properties:
        error:
          type: string
...
    FlatPagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_count:
          type: integer
          example: 5
...
    CartItemInput:
      type: object
      required: [cust_code, item_code, unit_code]
      properties:
        cust_code:
          type: string
          example: "C00001"
        emp_code:
          type: string
        guid_code:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        item_code:
          type: string
          example: "ITM001"
        item_name:
          type: string
        unit_code:
          type: string
          example: "ขวด"
        barcode:
          type: string
          default: ""
        qty:
          type: number
          default: 1
        price:
          type: number
          default: 0
        item_type:
          type: string
          enum: ["0", "3"]
          default: "0"
        wh_code:
          type: string
        shelf_code:
          type: string
        stand_value:
          type: number
          default: 1
        divide_value:
          type: number
          default: 1
        ratio:
          type: number
          default: 1
    CartItem:
      type: object
      properties:
        cust_code:
          type: string
        guid_code:
          type: string
          format: uuid
        item_code:
          type: string
        item_name:
          type: string
        unit_code:
          type: string
        item_type:
          type: string
        barcode:
          type: string
        qty:
          type: string
        price:
          type: string
        wh_code:
          type: string
        shelf_code:
          type: string
        creator_code:
          type: string
        create_datetime:
          type: string
          format: date-time
        stand_value:
          type: string
        divide_value:
          type: string
        ratio:
          type: string
        balance_qty:
          type: integer
          description: Always 0 — use /validatecartstock for real stock
    CartOrderItem:
      allOf:
        - $ref: '#/components/schemas/CartItem'
        - type: object
          properties:
            tax_type:
              type: integer
            price_confirm:
              type: string
              description: '"0" if item_type != "3"; else set price'
    StockCheckInput:
      type: object
      properties:
        item_code:
          type: string
        unit_code:
          type: string
    StockResult:
      type: object
      properties:
        item_code:
          type: string
        unit_code:
          type: string
        balance_qty:
          type: integer
...
    CartOrderPriceResult:
      type: object
      properties:
        item_code:
          type: string
        unit_code:
          type: string
        price_confirm:
          type: number
        success:
          type: boolean
        error_type:
          type: string
        message:
          type: string
...
paths:
...
  /additemtocart:
    post:
      tags: [Cart]
      summary: เพิ่มสินค้าลงตะกร้า
      description: |
        รับ **JSON array** โดยตรง (ไม่ใช่ object)

        แต่ละ item → **DELETE เดิมก่อน แล้ว INSERT ใหม่** (key: item_code + unit_code + barcode + cust_code)

        qty ที่ส่งมาจะ **แทนที่** ค่าเดิม ไม่ใช่บวกเพิ่ม
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/CartItemInput' }
            example:
              - cust_code: "C00001"
                emp_code: "EMP001"
                guid_code: "550e8400-e29b-41d4-a716-446655440000"
                item_code: "ITM001"
                item_name: "น้ำดื่ม ABC 600ml"
                unit_code: "ขวด"
                barcode: ""
                qty: 6
                price: 12.00
                item_type: "0"
                wh_code: "WH01"
                shelf_code: "A01"
                stand_value: 1
                divide_value: 1
                ratio: 1
      responses:
        '200':
          description: เพิ่มสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  msg: { type: string, example: "success" }
        '400':
          description: Body ไม่ใช่ JSON array หรือ database error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getcartitemlist:
    get:
      tags: [Cart]
      summary: รายการสินค้าในตะกร้า
      description: "`balance_qty` จะเป็น 0 เสมอ — ใช้ `/validatecartstock` สำหรับสต็อกจริง"
      parameters:
        - { name: cust_code,  in: query, required: true, schema: { type: string },                example: "C00001" }
        - { name: page,       in: query,                 schema: { type: integer, default: 1 },   description: "หน้าที่ (1-based)" }
        - { name: page_size,  in: query,                 schema: { type: integer, default: 20 },  description: "ต่อหน้า (max 200)" }
        - { name: search,     in: query,                 schema: { type: string },                description: "ค้นหา item_code หรือ item_name" }
      responses:
        '200':
          description: รายการในตะกร้า
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/FlatPagination'
                  - type: object
                    properties:
                      success: { type: boolean }
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CartItem' }
        '400':
          description: cust_code required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
  /getCartSummary:
    get:
      tags: [Cart]
      summary: สรุปยอดตะกร้า (ไม่คำนวณราคาใหม่)
      description: |
        คำนวณจาก qty × price ที่บันทึกไว้ใน DB โดยตรง — ไม่เรียก priceHelper

        ใช้ `/getcartfinalsummary` ถ้าต้องการราคาที่ถูกต้อง (แต่ช้ากว่า)
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: สรุปยอด
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  total_price: { type: number, example: 1250.50 }
                  total_qty:   { type: number, example: 18 }
                  total_items: { type: integer, example: 3 }
  /getcartitemstock:
    post:
      tags: [Cart]
      summary: ตรวจสต็อกสินค้าหลายรายการ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/StockCheckInput' }
            example:
              items:
                - { item_code: "ITM001", unit_code: "ขวด" }
                - { item_code: "ITM002", unit_code: "แพ็ค" }
      responses:
        '200':
          description: ยอดสต็อกต่อ item/unit
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/StockResult' }
        '400':
          description: Parse error (plain text หรือ JSON string)
  /getcartorder:
    get:
      tags: [Cart]
      summary: รายการตะกร้าพร้อม tax_type และ price_confirm
      description: |
        เหมือน getcartitemlist แต่ join กับ ic_inventory เพื่อดึง `tax_type`

        `price_confirm` = "0" ถ้า item_type ≠ "3" — ใช้ `/getcartorderprice` สำหรับราคาจริง
      parameters:
        - { name: cust_code,  in: query, required: true, schema: { type: string },               example: "C00001" }
        - { name: page,       in: query,                 schema: { type: integer, default: 1 } }
        - { name: page_size,  in: query,                 schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: รายการตะกร้า
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/FlatPagination'
                  - type: object
                    properties:
                      success: { type: boolean }
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CartOrderItem' }
  /getcartorderprice:
    post:
      tags: [Cart]
      summary: คำนวณราคาจริงของสินค้าในตะกร้า
      description: |
        เรียก getProductPriceLocalx() ต่อรายการ

        item_type = "3" (Set) → ใช้ `price` ที่ส่งมาโดยตรง

        ถ้า priceHelper fail → item นั้น return `success: false` (request ทั้งหมดไม่ fail)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cust_code, items]
              properties:
                cust_code: { type: string, example: "C00001" }
                items:
                  type: array
                  items: { $ref: '#/components/schemas/CartOrderPriceInput' }
      responses:
        '200':
          description: ราคาจริงต่อรายการ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CartOrderPriceResult' }
        '400':
          description: Parse error
...
  /getcartfinalsummary:
    get:
      tags: [Cart]
      summary: สรุปยอดรวมหลังคำนวณราคาจริง
      description: |
        Loop ทุก item แล้วเรียก getProductPriceLocalx() — **ช้าถ้าตะกร้ามีหลาย item**

        ต่างจาก `getCartSummary`: ได้ราคาที่ถูกต้อง แต่ใช้เวลามากกว่า
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: สรุปยอด
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      total_items: { type: integer }
                      total_qty:   { type: number }
                      total_price: { type: number }
  /validatecartstock:
    get:
      tags: [Cart]
      summary: ตรวจสอบสต็อกก่อนสั่งซื้อ
      description: รองรับทั้งสินค้าปกติและสินค้าชุด (Set) — ใช้ก่อน `/sendorder` เสมอ
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ผลการตรวจสต็อก
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:      { type: boolean }
                  is_valid:     { type: boolean, description: "true = ทุกรายการพอ" }
                  stock_issues:
                    type: array
                    items: { $ref: '#/components/schemas/StockIssue' }
              examples:
                valid:
                  value: { success: true, is_valid: true, stock_issues: [] }
                invalid:
                  value:
                    success: true
                    is_valid: false
                    stock_issues:
                      - { item_code: "ITM001", item_name: "น้ำดื่ม ABC", qty_in_cart: 100, balance_qty: 50, issue_type: "exceeding" }
        '400':
          description: cust_code required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponseLower' }
  /deleteItem:
    get:
      tags: [Cart]
      summary: ลบสินค้ารายการเดียวออกจากตะกร้า
      description: ลบด้วย guid_code + cust_code — ถ้าไม่พบ record ก็ return success:true
      parameters:
        - { name: guid_code,  in: query, required: true, schema: { type: string, format: uuid }, example: "550e8400-e29b-41d4-a716-446655440000" }
        - { name: cust_code,  in: query, required: true, schema: { type: string },               example: "C00001" }
      responses:
        '200':
          description: ลบสำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessOnly' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /deleteAllItems:
    get:
      tags: [Cart]
      summary: ล้างตะกร้าทั้งหมด
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ล้างสำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessOnly' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─── ORDER ───────────────────────────────────────────────────────────────────
...